refactor: final pass of edits to get things across the line as a bumped major version
