+++ src/index.ts
@@ -10,5 +10,3 @@ export { default as TextArea } from './components/TextArea';
 export { default as Typography } from './components/Typography';
 // utility exports
 export { default as paletteGenerator } from './utils/paletteGenerator';
-// plugin export
-export { default as skogen } from './lib/plugin';
