+++ styles/index.css
@@ -1,2 +1,2 @@
-@import 'tailwindcss';
 @import './fonts.css';
+@source "../dist";
