+++ tsdown.config.js
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/plugin.ts'],
   format: ['cjs', 'esm'],
   dts: true,
   sourcemap: true,
